requirements.txt
Django==3.2
psycopg2>=2.8.6,<2.9
asgiref==3.5.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
